Move logging infrastructure from the base_application to the example.
